feat#97: 이미지 생성 이력 결과 페이지 제공 메서드 생성
